+++ test/test_rotation.py
@@ -1,8 +1,6 @@
 from pathlib import Path
 
-import pytest
-
-from fpdf import FPDF, FPDFException
+from fpdf import FPDF
 from test.conftest import assert_pdf_equal
 
 HERE = Path(__file__).resolve().parent
@@ -17,25 +15,3 @@ def test_rotation(tmp_path):
         pdf.image(img_filepath, x=x, y=y)
     pdf.image(img_filepath, x=150, y=150)
     assert_pdf_equal(pdf, HERE / "rotation.pdf", tmp_path)
-
-
-def test_prevent_some_methods_in_rotation():  # issue-226
-    pdf = FPDF()
-    pdf.add_page()
-    with pdf.rotation(90):
-        with pytest.raises(FPDFException):
-            pdf.add_page()
-        with pytest.raises(FPDFException):
-            pdf.set_font("Times")
-        with pytest.raises(FPDFException):
-            pdf.set_font_size(16)
-        with pytest.raises(FPDFException):
-            pdf.set_draw_color(255)
-        with pytest.raises(FPDFException):
-            pdf.set_fill_color(255)
-        with pytest.raises(FPDFException):
-            pdf.set_text_color(255)
-        with pytest.raises(FPDFException):
-            pdf.set_line_width(2)
-        with pytest.raises(FPDFException):
-            pdf.set_stretching(10)