+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and [PEP 440](https://www.python.org/dev/peps/pep-0440/).
 
 ## [2.4.6] - not released yet
 ### Added
+- Temporary changes to graphics state variables are now possible by `with FPDF.local_context():`, thanks to @gmischler
 - a mechanism to detect & downscale oversized images,
   _cf._ [documentation](https://pyfpdf.github.io/fpdf2/Images.html#oversized-images-detection-downscaling).
   [Feedbacks](https://github.com/PyFPDF/fpdf2/discussions) on this new feature are welcome!
@@ -30,6 +31,7 @@ and [PEP 440](https://www.python.org/dev/peps/pep-0440/).
   It enables to draw solid arcs in a PDF document. A solid arc combines an arc and a triangle to form a pie slice.
 - [`FPDF.regular_polygon`](https://pyfpdf.github.io/fpdf2/fpdf/fpdf.html#fpdf.fpdf.FPDF.regular_polygon): new method added.
 ### Fixed
+- All graphics state manipulations are now possible within a rotation context, thanks to @gmischler
 - The exception making the "x2" template field optional for barcode elements did not work correctly, fixed by @gmischler
 ### Changed
 - All template elements now have a transparent default background instead of white, thanks to @gmischler